Refactoring
- Split utils into modules
- Clean up imports
- Clean up documentation
