+++ datacube/model/__init__.py
@@ -8,18 +8,21 @@ import logging
 import math
 import warnings
 from collections import namedtuple, OrderedDict, Sequence
+from datetime import datetime
 from pathlib import Path
 from uuid import UUID
 
+import yaml
 from affine import Affine
 
 from datacube.compat import urlparse
-from datacube.utils import geometry, without_lineage_sources
-from datacube.utils import parse_time, cached_property, uri_to_local_path, intersects, schema_validated, DocReader
+from datacube.utils import geometry, without_lineage_sources, parse_time, cached_property, uri_to_local_path, \
+    schema_validated, DocReader
 from datacube.utils.geometry import (CRS as _CRS,
                                      GeoBox as _GeoBox,
                                      Coordinate as _Coordinate,
-                                     BoundingBox as _BoundingBox)
+                                     BoundingBox as _BoundingBox, intersects)
+from datacube.utils.serialise import SafeDatacubeDumper
 
 _LOG = logging.getLogger(__name__)
 
@@ -521,6 +524,27 @@ class DatasetType(object):
     def dataset_reader(self, dataset_doc):
         return self.metadata_type.dataset_reader(dataset_doc)
 
+    def to_dict(self):
+        """
+        Convert to a dictionary representation of the available fields
+
+        :rtype: dict
+        """
+        row = {
+            'id': self.id,
+            'name': self.name,
+            'description': self.definition['description'],
+        }
+        row.update(self.fields)
+        if self.grid_spec is not None:
+            row.update({
+                'crs': str(self.grid_spec.crs),
+                'spatial_dimensions': self.grid_spec.dimensions,
+                'tile_size': self.grid_spec.tile_size,
+                'resolution': self.grid_spec.resolution,
+            })
+        return row
+
     def __str__(self):
         return "DatasetType(name={name!r}, id_={id!r})".format(id=self.id, name=self.name)
 
@@ -767,3 +791,23 @@ def metadata_from_doc(doc):
     from .fields import get_dataset_fields
     MetadataType.validate(doc)
     return MetadataType(doc, get_dataset_fields(doc))
+
+
+def _range_representer(dumper, data):
+    # type: (yaml.Dumper, Range) -> Node
+    begin, end = data
+
+    # pyyaml doesn't output timestamps in flow style as timestamps(?)
+    if isinstance(begin, datetime):
+        begin = begin.isoformat()
+    if isinstance(end, datetime):
+        end = end.isoformat()
+
+    return dumper.represent_mapping(
+        yaml.resolver.BaseResolver.DEFAULT_MAPPING_TAG,
+        (('begin', begin), ('end', end)),
+        flow_style=True
+    )
+
+
+SafeDatacubeDumper.add_representer(Range, _range_representer)
