+++ datacube/model/utils.py
@@ -14,7 +14,8 @@ from pandas import to_datetime
 
 import datacube
 from datacube.model import Dataset
-from datacube.utils import geometry, SimpleDocNav, sorted_items, InvalidDocException
+from datacube.utils import geometry, SimpleDocNav, InvalidDocException
+from datacube.utils.py import sorted_items
 
 try:
     from yaml import CSafeDumper as SafeDumper
