+++ datacube/storage/storage.py
@@ -21,7 +21,8 @@ from datacube.config import OPTIONS
 from datacube.drivers.datasource import DataSource
 from datacube.model import Dataset
 from datacube.storage import netcdf_writer
-from datacube.utils import clamp, datetime_to_seconds_since_1970, DatacubeException, ignore_exceptions_if
+from datacube.utils import datetime_to_seconds_since_1970, DatacubeException, ignore_exceptions_if
+from datacube.utils.math import clamp
 from datacube.utils import geometry
 from datacube.utils import is_url, uri_to_local_path, get_part_from_uri
 
@@ -512,7 +513,7 @@ def _url2rasterio(url_str, fmt, layer):
     url = urlparse(url_str)
     assert url.scheme, "Expecting URL with scheme here"
 
-    # if format is NETCDF of HDF need to pass NETCDF:path:band as filename to rasterio/GDAL
+    # if format is NETCDF or HDF need to pass NETCDF:path:band as filename to rasterio/GDAL
     for nasty_format in ('netcdf', 'hdf'):
         if nasty_format in fmt.lower():
             if url.scheme != 'file':
@@ -546,10 +547,12 @@ def _choose_location(dataset):
 
 
 def measurement_paths(dataset):
-    '''Returns a dictionary mapping from band name to url pointing to band storage
+    """
+    Returns a dictionary mapping from band name to url pointing to band storage
     resource.
+
     :return: {str: str} Band Name => URL
-    '''
+    """
     base = _choose_location(dataset)
     return dict((k, _resolve_url(base, m.get('path', '')))
                 for k, m in dataset.measurements.items())
