+++ datacube/ui/__init__.py
@@ -1,13 +1,12 @@
 """
 User Interface Utilities
 """
-from __future__ import absolute_import
-
 from .expression import parse_expressions
 from .common import get_metadata_path
 from datacube.utils import read_documents
 
 __all__ = [
     'parse_expressions',
-    'get_metadata_path'
+    'get_metadata_path',
+    "read_documents",
 ]
