+++ datacube/ui/click.py
@@ -77,8 +77,6 @@ class ClickHandler(logging.Handler):
         try:
             msg = self.format(record)
             click.echo(msg, err=True)
-        except (KeyboardInterrupt, SystemExit):
-            raise
         except:  # pylint: disable=bare-except
             self.handleError(record)
