+++ datacube/utils/geometry/__init__.py
@@ -8,6 +8,7 @@ from ._base import (
     CRS,
     Geometry,
     GeoBox,
+    intersects,
     scaled_down_geobox,
     point,
     multipoint,
@@ -46,6 +47,7 @@ __all__ = [
     "CRS",
     "Geometry",
     "GeoBox",
+    "intersects",
     "point",
     "multipoint",
     "line",
