+++ datacube_apps/pixeldrill.py
@@ -29,8 +29,8 @@ import datacube
 
 try:
     import tkinter as tk
-    import tkinter.ttk as ttk
-    import tkinter.font as font
+    from tkinter import tkk
+    from tkinter import font
 except ImportError:  # Python 2
     import Tkinter as tk
     import ttk
