+++ requirements-test.txt
@@ -3,14 +3,16 @@ amqp==2.2.2
 astroid==1.6.3
 attrs==17.4.0
 billiard==3.5.0.3
-boto3==1.4.3
-botocore==1.4.93
+boto==2.49.0
+boto3==1.7.80
+botocore==1.10.80
 cachetools==2.0.1
 celery==4.1.0
 certifi==2018.1.18
 cf-units==1.2.0
 cffi==1.11.5
 chardet==3.0.4
+ciso8601==2.1.1
 click==6.7
 click-plugins==1.0.3
 cligj==0.4.0
